Add files via upload
The first draft of the ports and cargo data for NOEP. This data was harvested from the US Census Bureau's Import/Export data. It's not complete, but the USACE Ports data is very distinct and will be presented in a different data set. Ports were ranked based on the average of their total values and total tonnage over the lifespan of the data.