Add files via upload
A list of Census Ports mapped to their counties and states. Census Ports are actually port districts. In fact, they are US customs districts. These may or may not be centered on the port itself. However, they will at least be in the same county. With one exception, Savannah in Georgia. The port district is in Effingham County which is immediately adjacent to the port.